Abstracting secrets provider, allowing in-mem secrets for integration tests.
