+++ mlrun/api/crud/secrets.py
@@ -23,11 +23,14 @@ import mlrun.api.utils.events.events_factory as events_factory
 import mlrun.api.utils.singletons.k8s
 import mlrun.common
 import mlrun.common.schemas
+import mlrun.common.secrets
 import mlrun.errors
 import mlrun.utils.helpers
 import mlrun.utils.regex
 import mlrun.utils.singleton
 import mlrun.utils.vault
+from mlrun.config import config as mlconf
+from mlrun.utils import logger
 
 
 class SecretsClientType(str, enum.Enum):
@@ -45,6 +48,24 @@ class Secrets(
     internal_secrets_key_prefix = "mlrun."
     # make it a subset of internal since key map are by definition internal
     key_map_secrets_key_prefix = f"{internal_secrets_key_prefix}map."
+
+    def __init__(self):
+        if mlconf.secret_stores.test_mode_mock_secrets:
+            logger.warning("***** USING SECRETS IN TEST MODE *****")
+            logger.warning(
+                "***** Secrets are kept in-memory. Only use this mode for testing *****"
+            )
+            self.secrets_provider = mlrun.common.secrets.InMemorySecretProvider()
+        else:
+            self.secrets_provider = mlrun.api.utils.singletons.k8s.get_k8s_helper()
+
+    @property
+    def secrets_provider(self) -> mlrun.common.secrets.SecretProviderInterface:
+        return self._secrets_provider
+
+    @secrets_provider.setter
+    def secrets_provider(self, provider: mlrun.common.secrets.SecretProviderInterface):
+        self._secrets_provider = provider
 
     def generate_client_project_secret_key(
         self, client_type: SecretsClientType, name: str, subtype=None
@@ -107,11 +128,8 @@ class Secrets(
             if secrets_to_store:
                 mlrun.utils.vault.store_vault_project_secrets(project, secrets_to_store)
         elif secrets.provider == mlrun.common.schemas.SecretProviderName.kubernetes:
-            if mlrun.api.utils.singletons.k8s.get_k8s_helper():
-                (
-                    secret_name,
-                    action,
-                ) = mlrun.api.utils.singletons.k8s.get_k8s_helper().store_project_secrets(
+            if self.secrets_provider:
+                (secret_name, action,) = self.secrets_provider.store_project_secrets(
                     project, secrets_to_store
                 )
                 secret_keys = [secret_name for secret_name in secrets_to_store.keys()]
@@ -138,10 +156,7 @@ class Secrets(
     def read_auth_secret(
         self, secret_name, raise_on_not_found=False
     ) -> mlrun.common.schemas.AuthSecretData:
-        (
-            username,
-            access_key,
-        ) = mlrun.api.utils.singletons.k8s.get_k8s_helper().read_auth_secret(
+        (username, access_key,) = self.secrets_provider.read_auth_secret(
             secret_name, raise_on_not_found=raise_on_not_found
         )
         return mlrun.common.schemas.AuthSecretData(
@@ -158,7 +173,7 @@ class Secrets(
             raise mlrun.errors.MLRunInvalidArgumentError(
                 f"Storing auth secret is not implemented for provider {secret.provider}"
             )
-        if not mlrun.api.utils.singletons.k8s.get_k8s_helper():
+        if not self.secrets_provider:
             raise mlrun.errors.MLRunInternalServerError(
                 "K8s provider cannot be initialized"
             )
@@ -167,9 +182,7 @@ class Secrets(
         (
             auth_secret_name,
             _,
-        ) = mlrun.api.utils.singletons.k8s.get_k8s_helper().store_auth_secret(
-            secret.username, secret.access_key
-        )
+        ) = self.secrets_provider.store_auth_secret(secret.username, secret.access_key)
 
         return auth_secret_name
 
@@ -182,11 +195,11 @@ class Secrets(
             raise mlrun.errors.MLRunInvalidArgumentError(
                 f"Storing auth secret is not implemented for provider {provider}"
             )
-        if not mlrun.api.utils.singletons.k8s.get_k8s_helper():
+        if not self.secrets_provider:
             raise mlrun.errors.MLRunInternalServerError(
                 "K8s provider cannot be initialized"
             )
-        mlrun.api.utils.singletons.k8s.get_k8s_helper().delete_auth_secret(secret_name)
+        self.secrets_provider.delete_auth_secret(secret_name)
 
     def delete_project_secrets(
         self,
@@ -218,13 +231,11 @@ class Secrets(
                 f"Delete secret is not implemented for provider {provider}"
             )
         elif provider == mlrun.common.schemas.SecretProviderName.kubernetes:
-            if mlrun.api.utils.singletons.k8s.get_k8s_helper():
+            if self.secrets_provider:
                 (
                     secret_name,
                     action,
-                ) = mlrun.api.utils.singletons.k8s.get_k8s_helper().delete_project_secrets(
-                    project, secrets
-                )
+                ) = self.secrets_provider.delete_project_secrets(project, secrets)
 
                 if action:
                     events_client = events_factory.EventsFactory().get_events_client()
@@ -267,12 +278,9 @@ class Secrets(
                     "Cannot specify token when requesting k8s secret keys"
                 )
 
-            if mlrun.api.utils.singletons.k8s.get_k8s_helper():
+            if self.secrets_provider:
                 secret_keys = (
-                    mlrun.api.utils.singletons.k8s.get_k8s_helper().get_project_secret_keys(
-                        project
-                    )
-                    or []
+                    self.secrets_provider.get_project_secret_keys(project) or []
                 )
             else:
                 raise mlrun.errors.MLRunInternalServerError(
@@ -316,10 +324,8 @@ class Secrets(
                 raise mlrun.errors.MLRunAccessDeniedError(
                     "Not allowed to list secrets data from kubernetes provider"
                 )
-            secrets_data = (
-                mlrun.api.utils.singletons.k8s.get_k8s_helper().get_project_secret_data(
-                    project, secrets
-                )
+            secrets_data = self.secrets_provider.get_project_secret_data(
+                project, secrets
             )
 
         else:
