+++ tests/api/conftest.py
@@ -34,6 +34,7 @@ import mlrun.api.utils.singletons.logs_dir
 import mlrun.api.utils.singletons.project_member
 import mlrun.api.utils.singletons.scheduler
 import mlrun.common.schemas
+import mlrun.common.secrets
 import mlrun.db.factory
 import mlrun.launcher.factory
 from mlrun import mlconf
@@ -142,10 +143,10 @@ async def async_client(db) -> Generator:
             yield async_client
 
 
-class K8sSecretsMock:
+class K8sSecretsMock(mlrun.common.secrets.InMemorySecretProvider):
     def __init__(self):
+        super().__init__()
         self._is_running_in_k8s = True
-        self.reset_mock()
 
     def reset_mock(self):
         # project -> secret_key -> secret_value
@@ -161,87 +162,6 @@ class K8sSecretsMock:
 
     def set_is_running_in_k8s_cluster(self, value: bool):
         self._is_running_in_k8s = value
-
-    @staticmethod
-    def get_auth_secret_name(username: str, access_key: str) -> str:
-        return f"secret-ref-{username}-{access_key}"
-
-    def store_auth_secret(
-        self, username: str, access_key: str, namespace=""
-    ) -> (str, mlrun.common.schemas.SecretEventActions):
-        secret_ref = self.get_auth_secret_name(username, access_key)
-        self.auth_secrets_map.setdefault(secret_ref, {}).update(
-            self._generate_auth_secret_data(username, access_key)
-        )
-        return secret_ref, mlrun.common.schemas.SecretEventActions.created
-
-    @staticmethod
-    def _generate_auth_secret_data(username: str, access_key: str):
-        return {
-            mlrun.common.schemas.AuthSecretData.get_field_secret_key(
-                "username"
-            ): username,
-            mlrun.common.schemas.AuthSecretData.get_field_secret_key(
-                "access_key"
-            ): access_key,
-        }
-
-    def delete_auth_secret(self, secret_ref: str, namespace=""):
-        del self.auth_secrets_map[secret_ref]
-
-    def read_auth_secret(self, secret_name, namespace="", raise_on_not_found=False):
-        secret = self.auth_secrets_map.get(secret_name)
-        if not secret:
-            if raise_on_not_found:
-                raise mlrun.errors.MLRunNotFoundError(
-                    f"Secret '{secret_name}' was not found in auth secrets map"
-                )
-
-            return None, None
-        username = secret[
-            mlrun.common.schemas.AuthSecretData.get_field_secret_key("username")
-        ]
-        access_key = secret[
-            mlrun.common.schemas.AuthSecretData.get_field_secret_key("access_key")
-        ]
-        return username, access_key
-
-    def store_project_secrets(
-        self, project, secrets, namespace=""
-    ) -> (str, mlrun.common.schemas.SecretEventActions):
-        self.project_secrets_map.setdefault(project, {}).update(secrets)
-        secret_name = project
-        return secret_name, mlrun.common.schemas.SecretEventActions.created
-
-    def delete_project_secrets(self, project, secrets, namespace=""):
-        if not secrets:
-            self.project_secrets_map.pop(project, None)
-        else:
-            for key in secrets:
-                self.project_secrets_map.get(project, {}).pop(key, None)
-        return "", True
-
-    def get_project_secret_keys(self, project, namespace="", filter_internal=False):
-        secret_keys = list(self.project_secrets_map.get(project, {}).keys())
-        if filter_internal:
-            secret_keys = list(
-                filter(lambda key: not key.startswith("mlrun."), secret_keys)
-            )
-        return secret_keys
-
-    def get_project_secret_data(self, project, secret_keys=None, namespace=""):
-        secrets_data = self.project_secrets_map.get(project, {})
-        return {
-            key: value
-            for key, value in secrets_data.items()
-            if (secret_keys and key in secret_keys) or not secret_keys
-        }
-
-    def store_secret(self, secret_name, secrets: dict):
-        self.secrets_map[secret_name] = secrets
-
-    def get_secret_data(self, secret_name, namespace=""):
-        return self.secrets_map[secret_name]
 
     def get_expected_env_variables_from_secrets(
         self, project, encode_key_names=True, include_internal=False, global_secret=None
@@ -313,32 +233,34 @@ class K8sSecretsMock:
             ] = ",".join(allowed_service_accounts)
         self.store_project_secrets(project, secrets)
 
+    def mock_functions(self, mocked_object, monkeypatch):
+        mocked_function_names = [
+            "is_running_inside_kubernetes_cluster",
+            "get_project_secret_keys",
+            "get_project_secret_data",
+            "store_project_secrets",
+            "delete_project_secrets",
+            "store_auth_secret",
+            "delete_auth_secret",
+            "read_auth_secret",
+            "get_secret_data",
+        ]
+
+        for mocked_function_name in mocked_function_names:
+            monkeypatch.setattr(
+                mocked_object,
+                mocked_function_name,
+                getattr(self, mocked_function_name),
+            )
+
 
 @pytest.fixture()
 def k8s_secrets_mock(monkeypatch, client: TestClient) -> K8sSecretsMock:
     logger.info("Creating k8s secrets mock")
     k8s_secrets_mock = K8sSecretsMock()
-
-    mocked_function_names = [
-        "is_running_inside_kubernetes_cluster",
-        "get_project_secret_keys",
-        "get_project_secret_data",
-        "store_project_secrets",
-        "delete_project_secrets",
-        "get_auth_secret_name",
-        "store_auth_secret",
-        "delete_auth_secret",
-        "read_auth_secret",
-        "get_secret_data",
-    ]
-
-    for mocked_function_name in mocked_function_names:
-        monkeypatch.setattr(
-            mlrun.api.utils.singletons.k8s.get_k8s_helper(),
-            mocked_function_name,
-            getattr(k8s_secrets_mock, mocked_function_name),
-        )
-
+    k8s_secrets_mock.mock_functions(
+        mlrun.api.utils.singletons.k8s.get_k8s_helper(), monkeypatch
+    )
     yield k8s_secrets_mock
